requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.10.0=py37_0
aiohttp=3.6.3=py37he774522_0
altair=4.1.0=pypi_0
argon2-cffi=20.1.0=pypi_0
asgiref=3.3.1=pypi_0
astor=0.8.1=py37_0
async-generator=1.10=pypi_0
async-timeout=3.0.1=py37_0
attrs=20.2.0=py_0
backcall=0.2.0=pypi_0
base58=2.0.1=pypi_0
beautifulsoup4=4.9.3=pypi_0
blas=1.0=mkl
bleach=3.2.1=pypi_0
blinker=1.4=py37_0
boto=2.49.0=py37_0
boto3=1.16.18=pypi_0
botocore=1.19.18=pypi_0
brotlipy=0.7.0=py37he774522_1000
bz2file=0.98=pypi_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=py_0
certifi=2020.11.8=py37haa95532_0
cffi=1.14.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
colorama=0.4.4=pypi_0
cryptography=3.1.1=py37h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py37_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
dill=0.3.2=py_0
django=3.1.3=pypi_0
django-bootstrap4=2.3.1=pypi_0
django-crispy-forms=1.10.0=pypi_0
docutils=0.16=py37_1
entrypoints=0.3=pypi_0
enum-compat=0.0.3=pypi_0
flask-dropzone=1.5.4=pypi_0
flask-wtf=0.14.3=pypi_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py37_1
gast=0.2.2=py37_0
gensim=3.8.0=py37hf9181ef_0
gitdb=4.0.5=pypi_0
gitpython=3.1.11=pypi_0
google-auth=1.22.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
googleapis-common-protos=1.52.0=py37h21ff451_0
grpcio=1.31.0=py37he7da953_0
h5py=2.8.0=py37hf7173ca_2
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
imagealign=1.0.0=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
imutils=0.5.3=pypi_0
intel-openmp=2020.2=254
ipykernel=5.3.4=pypi_0
ipython=7.18.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
jmespath=0.10.0=py_0
joblib=0.17.0=py_0
jpeg=9b=hb83a4c4_2
jsonpickle=1.4.1=py_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-console=6.2.0=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
keras=2.3.1=0
keras-applications=1.0.8=py_1
keras-base=2.3.1=py37_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37h74a9793_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0.1=h200bbdf_0
libtiff=4.1.0=h56a325e_1
llvmlite=0.31.0=py37ha925a31_0
lz4-c=1.9.2=hf4a77e7_3
markdown=3.3.2=py37_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37hba9282a_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
multidict=4.7.6=py37he774522_1
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.0.8=pypi_0
nest-asyncio=1.4.1=pypi_0
networkx=2.5=pypi_0
ninja=1.10.1=py37h7ef1ec2_0
nltk=3.5=pypi_0
notebook=6.1.4=pypi_0
numba=0.48.0=py37h47e9c7a_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
opencv-contrib-python=3.4.11.45=pypi_0
opencv-python=3.4.11.45=pypi_0
openssl=1.1.1h=he774522_0
opt_einsum=3.1.0=py_0
packaging=20.4=pypi_0
pandas=1.1.3=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.1=pypi_0
passlib=1.7.4=pypi_0
pathtools=0.1.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.0.0=py37hca74424_0
pip=20.2.4=py37_0
prometheus-client=0.8.0=pypi_0
promise=2.3=py37_0
prompt-toolkit=3.0.8=pypi_0
protobuf=3.13.0.1=py37ha925a31_1
psutil=5.7.2=py37he774522_0
py-opencv=3.4.2=py37hc319ecb_0
pyarrow=2.0.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pydeck=0.5.0=pypi_0
pygments=2.7.1=pypi_0
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
pytorch=1.6.0=py3.7_cuda101_cudnn7_0
pytz=2020.1=pypi_0
pywavelets=1.1.1=pypi_0
pywin32=228=pypi_0
pywinpty=0.5.7=pypi_0
pyyaml=5.3.1=py37he774522_1
pyzmq=19.0.2=pypi_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.7.7=pypi_0
qtpy=1.9.0=pypi_0
regex=2020.10.23=pypi_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
s3transfer=0.3.3=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.5.2=py37h9439919_0
seaborn=0.11.0=pypi_0
send2trash=1.5.0=pypi_0
setuptools=50.3.0=py37h9490d1a_1
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
smart-open=2.1.0=pypi_0
smmap=3.0.4=pypi_0
sortedcontainers=2.2.2=pypi_0
soupsieve=2.0.1=pypi_0
sqlite=3.33.0=h2a8f88b_0
sqlparse=0.4.1=pypi_0
streamlit=0.71.0=pypi_0
tbb=2020.3=h74a9793_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.1.0=gpu_py37h7db9008_0
tensorflow-base=2.1.0=gpu_py37h55f5790_0
tensorflow-datasets=1.2.0=py37_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-gpu=2.1.0=h0d30ee6_0
tensorflow-hub=0.9.0=pypi_0
tensorflow-metadata=0.14.0=pyhe6710b0_1
termcolor=1.1.0=py37_1
terminado=0.9.1=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=pypi_0
tk=0.1.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.11.1=pypi_0
torchvision=0.7.0=py37_cu101
tornado=6.0.4=py37he774522_1
tqdm=4.50.2=py_0
traitlets=5.0.5=pypi_0
tzlocal=2.1=pypi_0
urllib3=1.24.3=py37_0
validators=0.18.1=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
watchdog=0.10.3=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=0.16.1=py_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=pypi_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
wrapt=1.12.1=py37he774522_1
wtforms=2.3.3=pypi_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yarl=1.6.2=py37he774522_0
zipp=3.3.1=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
